+++ libs/naas-abi-core/CHANGELOG.md
@@ -2,6 +2,31 @@
 
 <!-- version list -->
 
+## v1.7.1 (2026-01-26)
+
+### Bug Fixes
+
+- CICD checks
+  ([`1cc9a0d`](https://github.com/jupyter-naas/abi/commit/1cc9a0d8aa9e5b89c06269f11b8437d3a60d2c23))
+
+
+## v1.7.0 (2026-01-24)
+
+
+## v1.6.0 (2026-01-24)
+
+### Features
+
+- Working on improving project initialization
+  ([`e7063bb`](https://github.com/jupyter-naas/abi/commit/e7063bb63a7d3b553f159861fb3e1c545e7d8849))
+
+
+## v1.5.1 (2026-01-16)
+
+
+## v1.5.0 (2026-01-15)
+
+
 ## v1.4.2 (2026-01-08)
 
 ### Bug Fixes